+++ lib/pgchief/command.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Pgchief
+  # Base module for command classes
+  module Command; end
+end