+++ lib/pgchief/prompt.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Pgchief
+  # Base module for prompt classes
+  module Prompt; end
+end