+++ lib/pgchief/prompt/create_user.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Pgchief
+  module Prompt
+    # Class to prompt for user creation details
+    class CreateUser < Base
+      def self.call
+        prompt   = TTY::Prompt.new
+        username = prompt.ask("Username:")
+        password = prompt.mask("Password:")
+        result   = Pgchief::Command::UserCreate.call(username, password)
+
+        prompt.say result
+      end
+    end
+  end
+end