+++ lib/pgchief/prompt/drop_database.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Pgchief
+  module Prompt
+    # Class to prompt for which database to drop
+    class DropDatabase < Base
+      def self.call
+        prompt   = TTY::Prompt.new
+        database = prompt.select("Which database needs to be dropped?", Pgchief::Database.all)
+        result   = Pgchief::Command::DatabaseDrop.call(database)
+
+        prompt.say result
+      end
+    end
+  end
+end