+++ lib/pgchief/user.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require "pg"
+
+module Pgchief
+  # Database information and operations
+  class User
+    def self.all
+      conn   = PG.connect(ENV.fetch("DATABASE_URL"))
+      result = conn.exec("SELECT usename FROM pg_user")
+
+      result.map { |row| row["usename"] }.reject { |name| name == "postgres" }
+    ensure
+      conn.close
+    end
+  end
+end