+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## [0.1.0] - 2024-07-20
 
 - Initial release
+- Create database ✅
+- Create user ✅
+- Drop database ✅
+- Drop user ✅
+- List databases ✅